Rename gem name, 'require' name
