+++ Rakefile
@@ -24,7 +24,7 @@ TESTDIR = BASEDIR + "tmp_test_*"
 DLEXT   = RbConfig::CONFIG['DLEXT']
 EXT     = LIBDIR + "pg_ext.#{DLEXT}"
 
-GEMSPEC = 'yugabyte_ysql.gemspec'
+GEMSPEC = 'yugabytedb-ysql.gemspec'
 
 CLEAN.include( TESTDIR.to_s )
 CLEAN.include( PKGDIR.to_s, TMPDIR.to_s )
