+++ lib/pg/basic_type_map_based_on_result.rb
@@ -1,7 +1,7 @@
 # -*- ruby -*-
 # frozen_string_literal: true
 
-require 'yugabyte_ysql' unless defined?( YugabyteYSQL )
+require 'ysql' unless defined?( YugabyteYSQL )
 
 # Simple set of rules for type casting common PostgreSQL types from Ruby
 # to PostgreSQL.
