+++ lib/pg/load_balance_service.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-require 'yugabyte_ysql' unless defined?( YugabyteYSQL )
+require 'ysql' unless defined?( YugabyteYSQL )
 require 'concurrent'
 
 class YugabyteYSQL::LoadBalanceService
