+++ lib/pg/type_map_by_column.rb
@@ -1,7 +1,7 @@
 # -*- ruby -*-
 # frozen_string_literal: true
 
-require 'yugabyte_ysql' unless defined?( YugabyteYSQL )
+require 'ysql' unless defined?( YugabyteYSQL )
 
 class YugabyteYSQL::TypeMapByColumn
 	# Returns the type oids of the assigned coders.
