+++ sample/test_binary_values.rb
@@ -1,6 +1,6 @@
 # -*- ruby -*-1.9.1
 
-require 'yugabyte_ysql'
+require 'ysql'
 
 db = YugabyteYSQL.connect(:dbname => 'test' )
 db.exec "DROP TABLE IF EXISTS test"
