+++ sample/warehouse_partitions.rb
@@ -41,7 +41,7 @@ require 'ostruct'
 require 'optparse'
 require 'pathname'
 require 'etc'
-require 'yugabyte_ysql'
+require 'ysql'
 
 
 ### A tablespace migration class.
