+++ spec/pg/basic_type_map_for_queries_spec.rb
@@ -3,7 +3,7 @@
 
 require_relative '../helpers'
 
-require 'yugabyte_ysql'
+require 'ysql'
 require 'time'
 
 describe 'Basic type mapping' do
