+++ spec/pg/type_map_by_column_spec.rb
@@ -3,7 +3,7 @@
 
 require_relative '../helpers'
 
-require 'yugabyte_ysql'
+require 'ysql'
 
 
 describe YugabyteYSQL::TypeMapByColumn do
